requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.8.3=pypi_0
aiosignal=1.3.1=pypi_0
asttokens=2.1.0=pypi_0
async-timeout=4.0.2=pypi_0
attrs=22.1.0=pypi_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py39h2bbff1b_1003
ca-certificates=2022.10.11=haa95532_0
certifi=2022.9.24=py39haa95532_0
cffi=1.15.1=py39h2bbff1b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=pypi_0
cryptography=38.0.1=py39h21b164f_0
cuda=11.7.1=0
cuda-cccl=11.7.91=0
cuda-command-line-tools=11.7.1=0
cuda-compiler=11.7.1=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cuobjdump=11.7.91=0
cuda-cupti=11.7.101=0
cuda-cuxxfilt=11.7.91=0
cuda-demo-suite=11.8.86=0
cuda-documentation=11.8.86=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-memcheck=11.8.86=0
cuda-nsight-compute=11.8.0=0
cuda-nvcc=11.7.99=0
cuda-nvdisasm=11.8.86=0
cuda-nvml-dev=11.7.91=0
cuda-nvprof=11.8.87=0
cuda-nvprune=11.7.91=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-nvvp=11.8.87=0
cuda-runtime=11.7.1=0
cuda-sanitizer-api=11.8.86=0
cuda-toolkit=11.7.1=0
cuda-tools=11.7.1=0
cuda-visual-tools=11.7.1=0
debugpy=1.6.3=pypi_0
decorator=5.1.1=pypi_0
entrypoints=0.4=pypi_0
executing=1.2.0=pypi_0
fire=0.4.0=pypi_0
freetype=2.12.1=ha860e81_0
frozenlist=1.3.3=pypi_0
fsspec=2022.11.0=pypi_0
idna=3.4=py39haa95532_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.17.1=pypi_0
ipython=8.6.0=pypi_0
jedi=0.18.2=pypi_0
jpeg=9e=h2bbff1b_0
jupyter-client=7.4.7=pypi_0
jupyter-core=5.0.0=pypi_0
lerc=3.0=hd77b12b_0
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.0.86=0
libcurand-dev=10.3.0.86=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.8=h2bbff1b_5
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.37=h2a8f88b_0
libtiff=4.4.0=h8a3f274_2
libuv=1.40.0=he774522_0
libwebp=1.2.4=h2bbff1b_0
libwebp-base=1.2.4=h2bbff1b_0
lightning-utilities=0.3.0=pypi_0
lz4-c=1.9.3=h2bbff1b_1
matplotlib-inline=0.1.6=pypi_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
multidict=6.0.2=pypi_0
nest-asyncio=1.5.6=pypi_0
nsight-compute=2022.3.0.22=0
numpy=1.23.4=py39h3b20f71_0
numpy-base=1.23.4=py39h4da318b_0
openssl=1.1.1s=h2bbff1b_0
parso=0.8.3=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.2.0=py39hdc2b20a_1
pip=22.2.2=py39haa95532_0
platformdirs=2.5.4=pypi_0
prompt-toolkit=3.0.33=pypi_0
protobuf=3.20.1=pypi_0
psutil=5.9.4=pypi_0
pure-eval=0.2.2=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.13.0=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pysocks=1.7.1=py39haa95532_0
python=3.9.15=h6244533_1
python-dateutil=2.8.2=pypi_0
pytorch=1.13.0=py3.9_cuda11.7_cudnn8_0
pytorch-cuda=11.7=h67b0de4_0
pytorch-lightning=1.8.3.post1=pypi_0
pytorch-mutex=1.0=cuda
pyzmq=24.0.1=pypi_0
requests=2.28.1=py39haa95532_0
setuptools=65.5.0=py39haa95532_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.40.0=h2bbff1b_0
stack-data=0.6.1=pypi_0
tensorboardx=2.5.1=pypi_0
termcolor=2.1.1=pypi_0
tk=8.6.12=h2bbff1b_0
torchaudio=0.13.0=pypi_0
torchmetrics=0.10.3=pypi_0
torchvision=0.14.0=pypi_0
tornado=6.2=pypi_0
tqdm=4.64.1=pypi_0
traitlets=5.5.0=pypi_0
typing_extensions=4.3.0=py39haa95532_0
tzdata=2022f=h04d1e81_0
urllib3=1.26.12=py39haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py39haa95532_0
wincertstore=0.2=py39haa95532_2
xz=5.2.6=h8cc25b3_0
yarl=1.8.1=pypi_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.2=h19a0ad4_0
